refactor(habitat): move baseline expansion into check module
