+++ tools/habitat-harness/test/lib/hooks.test.ts
@@ -25,7 +25,6 @@ import type {
   CheckOptions,
   CheckReport,
 } from "@internal/habitat-harness/service/model/check/index";
-import type { BaselineExpansionResult } from "@internal/habitat-harness/service/model/check/policy/structural/index";
 import {
   classifyResourcePreCommitDecisionEffect,
   classifyResourcesState,
@@ -39,14 +38,10 @@ import { makeTestHabitatServiceDeps } from "../support/habitat-service-deps";
 
 type StructuralCheckPolicy = {
   readonly createReport: (options?: CheckOptions) => Effect.Effect<CheckReport>;
-  readonly expandBaselines: () => Effect.Effect<BaselineExpansionResult>;
 };
 
 const mockCreateCheckReportEffect = vi.hoisted(() =>
   vi.fn<StructuralCheckPolicy["createReport"]>()
-);
-const mockExpandBaselinesEffect = vi.hoisted(() =>
-  vi.fn<StructuralCheckPolicy["expandBaselines"]>()
 );
 
 vi.mock(
@@ -59,7 +54,6 @@ vi.mock(
     return {
       ...actual,
       createCheckReportEffect: mockCreateCheckReportEffect,
-      expandBaselinesEffect: mockExpandBaselinesEffect,
     };
   }
 );
@@ -494,13 +488,11 @@ function makeStructuralCheckPolicy(fake: FakeHookHarness): StructuralCheckPolicy
         }
         return passingCheckReport(options.command?.serialized ?? "habitat check");
       }),
-    expandBaselines: () => Effect.succeed({ ok: true, messages: [] }),
   };
 }
 
 function useStructuralCheckPolicy(policy: StructuralCheckPolicy) {
   mockCreateCheckReportEffect.mockImplementation(policy.createReport);
-  mockExpandBaselinesEffect.mockImplementation(policy.expandBaselines);
 }
 
 function makeBiomeLayer(fake: FakeHookHarness) {
